+++ src/features/buy-now/BuyNowForm/FormSteps/index.ts
@@ -1,3 +1,2 @@
-export * from './ApproveZAuction';
 export * from './Complete';
 export * from './Details';
