--- src/features/buy-now/BuyNowForm/FormSteps/ui/ErrorText/ErrorText.tsx
@@ -1,9 +1,0 @@
-import styles from './ErrorText.module.scss';
-
-export interface ErrorTextProps {
-	text: string;
-}
-
-export const ErrorText = ({ text }: ErrorTextProps) => (
-	<span className={styles.Error}>{text}</span>
-);
