--- src/features/buy-now/BuyNowForm/FormSteps/ui/ErrorText/index.ts
@@ -1,1 +1,0 @@
-export * from './ErrorText';
