--- src/features/buy-now/BuyNowForm/FormSteps/ui/TextContent/TextContent.tsx
@@ -1,24 +1,0 @@
-import { ErrorText } from '../ErrorText/ErrorText';
-
-import styles from './TextContent.module.scss';
-
-export interface TextContentProps {
-	variant?: 'success';
-	errorText?: string;
-	textContent: string;
-}
-
-export const TextContent = ({
-	variant,
-	errorText,
-	textContent,
-}: TextContentProps) => {
-	return (
-		<>
-			<span className={styles.TextContent} data-variant={variant}>
-				{textContent}
-			</span>
-			{errorText !== undefined && <ErrorText text={errorText} />}
-		</>
-	);
-};
