--- src/features/buy-now/BuyNowForm/FormSteps/ui/TextContent/index.ts
@@ -1,1 +1,0 @@
-export * from './TextContent';
