+++ src/features/buy-now/BuyNowForm/FormSteps/ui/index.ts
@@ -1,3 +1,1 @@
-export * from './ErrorText';
 export * from './NFTDetails';
-export * from './TextContent';
