+++ src/features/ui/ApproveZAuction/ApproveZAuction.module.scss
@@ -1,0 +1,14 @@
+@import '../../../../node_modules/@zero-tech/zui/styles/_colors.scss';
+@import '../../../../node_modules/@zero-tech/zui/styles/_typography.scss';
+
+.Container {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	text-align: center;
+
+	.Confirmation {
+		margin-bottom: 1.125rem;
+		white-space: pre-line;
+	}
+}
