+++ src/features/ui/ApproveZAuction/ApproveZAuction.tsx
@@ -1,15 +1,17 @@
-import { TextContent } from '../ui';
+import { FormTextContent } from '../FormTextContent';
 import { Wizard } from '@zero-tech/zui/components/Wizard';
 
-import styles from '../FormSteps.module.scss';
+import styles from './ApproveZAuction.module.scss';
 
 export interface ApproveZAuctionProps {
+	action: string;
 	errorText: string;
 	onClose: () => void;
 	onApproveZAuction: () => void;
 }
 
 export const ApproveZAuction = ({
+	action,
 	errorText,
 	onClose,
 	onApproveZAuction,
@@ -17,10 +19,8 @@ export const ApproveZAuction = ({
 	const primaryButtonText = errorText ? 'Retry' : 'Continue';
 
 	const confirmationMessage = (
-		<TextContent
-			textContent={
-				'Before you can buy now, your wallet needs to approve zAuction. \nYou will only need to do this once. This will cost gas.'
-			}
+		<FormTextContent
+			textContent={`Before you can ${action}, your wallet needs to approve zAuction. \nYou will only need to do this once. This will cost gas.`}
 			errorText={errorText}
 		/>
 	);
