+++ src/features/ui/FormErrorText/FormErrorText.tsx
@@ -1,0 +1,9 @@
+import styles from './FormErrorText.module.scss';
+
+export interface FormErrorTextProps {
+	text: string;
+}
+
+export const FormErrorText = ({ text }: FormErrorTextProps) => (
+	<span className={styles.Error}>{text}</span>
+);
