+++ src/features/ui/FormErrorText/index.ts
@@ -1,0 +1,1 @@
+export * from './FormErrorText';
