+++ src/features/ui/FormTextContent/FormTextContent.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../../../../node_modules/@zero-tech/zui/styles/_colors.scss';
+@import '../../../../node_modules/@zero-tech/zui/styles/_colors.scss';
 
 .TextContent {
 	line-height: 1.5rem;
