+++ src/features/ui/FormUserBalance/FormUserBalance.module.scss
@@ -1,0 +1,17 @@
+@import '../../../../node_modules/@zero-tech/zui/styles/_colors.scss';
+@import '../../../../node_modules/@zero-tech/zui/styles/_typography.scss';
+
+.UserBalanceContainer {
+	display: flex;
+	flex-direction: column;
+
+	.Label {
+		color: $grey-lighter-2;
+		font-size: $size-tiny;
+		line-height: 0.908rem;
+	}
+
+	.Balance {
+		line-height: 1.5rem;
+	}
+}
