+++ src/features/ui/index.ts
@@ -1,6 +1,10 @@
+export * from './ApproveZAuction';
 export * from './ConnectWallet';
 export * from './FormDetails';
+export * from './FormErrorText';
 export * from './FormInputs';
+export * from './FormTextContent';
+export * from './FormUserBalance';
 export * from './IconButton';
 export * from './OptionLabel';
 export * from './StatsList';
